backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0

langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
langchain-core==0.1.10

chromadb==0.4.22
sentence-transformers>=2.3.1
numpy>=2.1.0
pandas>=2.2.3
openpyxl==3.1.2

PyPDF2==3.0.1
python-docx==1.1.0

aiosqlite==0.19.0

pydantic==2.5.3
pydantic-settings==2.1.0

cryptography==42.0.0
python-jose[cryptography]==3.3.0

torch>=2.1.0   # atau versi yang sudah mendukung 3.13
scikit-learn==1.4.0
